+++ lib/active_currency.rb
@@ -8,6 +8,7 @@ require 'active_currency/engine'
 require 'active_currency/database_store'
 require 'active_currency/cacheable_store'
 require 'active_currency/rate_store'
+require 'active_currency/memory_rate_store'
 require 'active_currency/add_rates'
 
 module ActiveCurrency